fix #49 : safari is legitimately cancer.
it turns out that safari
a) doesn't grow items inside a display: flex 'd container, and
b) doesn't recalculate flex when content changes.

this piece of shit browser should be put up to a wall and shot.
everyone who denies that it's this generations IE is either a liar or a retard.
the amount of inconsistencies to the spec/ other browsers is staggering.
who cares for 100% es6 compliance if fuckinf FLEXBOX doesn't work properly.

microsoft was at least nice enough to provide vm images for people who don't own a windows machine.
apple says fuck you, give me money to be able to fix shit.